requirements.txt
# API e Web
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
pydantic==2.4.2

# Testes
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
coverage==7.3.2
httpx==0.25.1

# Qualidade de Código
black==23.10.1
isort==5.12.0
mypy==1.6.1
flake8==7.0.0
flake8-docstrings==1.7.0
bandit==1.7.6
pre-commit==3.6.0

# Documentação
mkdocs==1.5.3
mkdocs-material==9.5.3
mkdocstrings==0.24.0 